Split of `stars.py` module
